[CE-3092] Highlight UPI instruments on desktop in p13n and run experiment (#2608)
* [CE-3092] Highlight UPI instruments on desktop in p13n and run experiment

Currently all the instruments with vpa ‘@ybl, @okhdfcbank etc are considered as UPI intent instruments as the first part (before @) of the vpa is not present, and we can identify the app with the second part (after @) of the vpa. But desktops can’t have intent instruments, so it is currently filtered out. This change shows the intent instrument on desktop but when clicked on it, it takes the user to upi collect screen and prefills the second part of the vpa (@ybl, @paytm etc.)

* Cleanup imports

* Add Tests

* Add Analytics

* Pass preferred_methods from the test

* Apply changes from review and fix analytics

* Apply changes from review

* import cleanup

* update experiments

* Fix upi app icon for IE 11